Extract distribution utils to `dwave.cloud.utils.dist`
