+++ dwave/cloud/cli.py
@@ -36,7 +36,7 @@ from dwave.cloud.utils.qubo import generate_random_ising_problem
 from dwave.cloud.utils.http import user_agent
 from dwave.cloud.utils.time import datetime_to_timestamp, utcnow, epochnow
 from dwave.cloud.utils.cli import default_text_input, strtrunc, CLIError
-from dwave.cloud.utils import (
+from dwave.cloud.utils.dist import (
     get_contrib_packages, get_distribution, PackageNotFoundError, VersionNotFoundError)
 from dwave.cloud.coders import bqm_as_file
 from dwave.cloud.package_info import __title__, __version__
