+++ tests/test_utils.py
@@ -40,10 +40,10 @@ from dwave.cloud.utils.qubo import (
 from dwave.cloud.utils.time import utcnow
 from dwave.cloud.utils.cli import default_text_input
 from dwave.cloud.utils.decorators import aliasdict, cached, deprecated, retried
+from dwave.cloud.utils.dist import get_distribution, PackageNotFoundError, VersionNotFoundError
 from dwave.cloud.utils import (
     NumpyEncoder, coerce_numpy_to_python,
-    hasinstance, exception_chain,
-    is_caused_by, get_distribution, PackageNotFoundError, VersionNotFoundError)
+    hasinstance, exception_chain, is_caused_by)
 
 logger = logging.getLogger(__name__)
 
@@ -315,19 +315,19 @@ class TestCachedCommon(unittest.TestCase):
             return next(counter)
 
         # populate
-        with mock.patch('dwave.cloud.utils.epochnow', lambda: 0):
+        with mock.patch('dwave.cloud.utils.time.epochnow', lambda: 0):
             self.assertEqual(f(), 0)
             self.assertEqual(f(1), 1)
             self.assertEqual(f(a=1, b=2), 2)
 
         # cache miss, expired
-        with mock.patch('dwave.cloud.utils.epochnow', lambda: 15):
+        with mock.patch('dwave.cloud.utils.time.epochnow', lambda: 15):
             self.assertEqual(f(), 3)
             self.assertEqual(f(1), 4)
             self.assertEqual(f(a=1, b=2), 5)
 
         # cache hit, still hot
-        with mock.patch('dwave.cloud.utils.epochnow', lambda: 5):
+        with mock.patch('dwave.cloud.utils.time.epochnow', lambda: 5):
             self.assertEqual(f(), 3)
             self.assertEqual(f(1), 4)
             self.assertEqual(f(a=1, b=2), 5)
@@ -340,17 +340,17 @@ class TestCachedCommon(unittest.TestCase):
             return next(counter)
 
         # populate
-        with mock.patch('dwave.cloud.utils.epochnow', lambda: 0):
+        with mock.patch('dwave.cloud.utils.time.epochnow', lambda: 0):
             self.assertEqual(f(), 0)
 
         # expired for default maxage
-        with mock.patch('dwave.cloud.utils.epochnow', lambda: 15):
+        with mock.patch('dwave.cloud.utils.time.epochnow', lambda: 15):
             self.assertEqual(f(maxage_=20), 0)
             self.assertEqual(f(maxage_=15.01), 0)
             self.assertEqual(f(maxage_=15), 1)
 
         # cache hot for default maxage
-        with mock.patch('dwave.cloud.utils.epochnow', lambda: 25):
+        with mock.patch('dwave.cloud.utils.time.epochnow', lambda: 25):
             self.assertEqual(f(maxage_=11), 1)
             self.assertEqual(f(), 2)
             self.assertEqual(f(maxage_=1), 2)
